move classes to own file
